Update README.md
Intro to Default Parameters

Step 1

Examine the code and try to predict how it will execute. After running the program, add comments to the top of the code explaining why the different function calls to displayStars produce the observed outputs.

Step 2

Adjust the program so that the displayStars function will produce 5 columns and 7 rows of stars when called using its default parameters. Rerun the program to ensure the default parameters are being.